+++ front/.prettierignore
@@ -1,10 +1,10 @@
 *.md
 public/locales/en/infraEditor.json
+src/applications/OperationalStudies/views/Scenario/components/ImportTimetableItem/rollingstock_opendata2osrd.json
 src/assets/mapstyles/*.json
 src/common/ReleaseInformations/json/licenses.json
 src/common/api/generatedEditoastApi.ts
 src/common/api/osrdGatewayApi.ts
-src/modules/timetableItem/components/ImportTimetableItem/rollingstock_opendata2osrd.json
 src/reducers/osrdconf/infra_schema.json
 src/utils/__tests__/assets/*.json
 /ui
