+++ front/src/applications/operationalStudies/views/Scenario/components/ImportTimetableItem/generateTrainSchedulesPayloads.ts
