+++ front/src/applications/operationalStudies/views/Scenario/components/ImportTimetableItem/helpers/buildStepsFromOcp.ts
