+++ front/src/applications/operationalStudies/views/Scenario/components/ImportTimetableItem/index.ts
@@ -1,0 +1,3 @@
+import ImportTimetableItem from './ImportTimetableItem';
+
+export default ImportTimetableItem;
