+++ front/src/applications/operationalStudies/views/Scenario/components/ImportTimetableItem/types.ts
@@ -1,0 +1,3 @@
+import type rollingstockOpenData2OSRD from './rollingstock_opendata2osrd.json';
+
+export type RollingstockOpenData2OSRDKeys = keyof typeof rollingstockOpenData2OSRD;
