--- front/src/modules/timetableItem/components/ImportTimetableItem/index.ts
@@ -1,3 +1,0 @@
-export { default as ImportTimetableItemConfig } from 'modules/timetableItem/components/ImportTimetableItem/ImportTimetableItemConfig';
-export { default as ImportTimetableItemStationSelector } from 'modules/timetableItem/components/ImportTimetableItem/ImportTimetableItemStationSelector';
-export { default as ImportTimetableItemTrainsList } from 'modules/timetableItem/components/ImportTimetableItem/ImportTimetableItemTrainsList';
