--- front/src/modules/timetableItem/components/ImportTimetableItem/types.ts
@@ -1,3 +1,0 @@
-import type rollingstockOpenData2OSRD from 'modules/timetableItem/components/ImportTimetableItem/rollingstock_opendata2osrd.json';
-
-export type RollingstockOpenData2OSRDKeys = keyof typeof rollingstockOpenData2OSRD;
